hrz-3074: confirmation banner refactoring
