+++ libs/domain/order/confirmation-banner/confirmation-banner.component.ts
@@ -1,7 +1,7 @@
 import { ContentMixin } from '@spryker-oryx/experience';
 import { OrderMixin } from '@spryker-oryx/order';
 import { HeadingTag } from '@spryker-oryx/ui/heading';
-import { asyncState, i18n, valueType } from '@spryker-oryx/utilities';
+import { i18n, signal } from '@spryker-oryx/utilities';
 import { html, LitElement, TemplateResult } from 'lit';
 import { styles } from './confirmation-banner.styles';
 
@@ -10,19 +10,22 @@ export class OrderConfirmationBannerComponent extends OrderMixin(
 ) {
   static styles = styles;
 
-  @asyncState()
-  protected orderRef = valueType(this.orderController.getRef());
+  protected orderRef = signal(this.orderController.getRef());
 
   protected override render(): TemplateResult {
-    return html`<oryx-heading .as=${HeadingTag.H2}>
+    return html`
+      <section>
         <oryx-image resource="order-confirmation-success"></oryx-image>
-        <h1>${i18n('order.confirmation.thank-you')}</h1>
-      </oryx-heading>
+        <oryx-heading .as=${HeadingTag.H2}>
+          <h1>${i18n('order.confirmation.thank-you')}</h1>
+        </oryx-heading>
+      </section>
       <p>
         ${i18n('order.confirmation.your-order-<id>-placed', {
           id: this.orderRef,
         })}
       </p>
-      <p>${i18n('order.confirmation.email-sent')}</p> `;
+      <p>${i18n('order.confirmation.email-sent')}</p>
+    `;
   }
 }
