+++ libs/domain/order/confirmation-banner/confirmation-banner.def.ts
