+++ libs/domain/order/confirmation-banner/confirmation-banner.schema.ts
