+++ libs/domain/order/confirmation-banner/index.ts
@@ -1,1 +1,3 @@
-export * from './src';
+export * from './confirmation-banner.component';
+export * from './confirmation-banner.schema';
+export * from './confirmation-banner.styles';
