--- libs/domain/order/confirmation-banner/src/index.ts
@@ -1,2 +1,0 @@
-export * from './confirmation-banner.component';
-export * from './confirmation-banner.styles';
