+++ libs/domain/order/src/components.ts
@@ -1,4 +1,4 @@
-export * from '../confirmation-banner/src/confirmation-banner.def';
+export * from '../confirmation-banner/confirmation-banner.def';
 export * from '../entries/src/entries.def';
 export * from '../summary/src/summary.def';
 export * from '../totals/src/totals.def';
